refactor: 북마크 상세 페이지 라우트 변경
