+++ src/components/pages/Places/bookmark/[bookmarkId]/BookmarkDetailList.tsx
